fix: balance type

+++ src/main/java/ht/henrique/mazebank/model/database/User.java
@@ -5,6 +5,9 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+import java.text.DecimalFormat;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -21,7 +24,7 @@ public class User implements Projection{
     @Column(name = "user_pass")
     private String user_pass;
     @Column(name = "user_balance")
-    private Integer user_balance;
+    private BigDecimal user_balance;
     @Column(name = "user_email")
     private String user_email;
 }
